# chore: improve performance of ClearInternal